Reworked TEXTMAP parsing
No more RegEx, that + the separate code comment pass was killing the load time.

Also, added a global for easily making performance timers.

Medium maps now load in a reasonable amount of time. Large maps still hang for a little bit due to the number of lines that need to be drawn all at once, but it's much more reasonable.